+++ Example/combo/index.web.js
@@ -1,0 +1,11 @@
+import React, { Component } from 'react';
+import { Text } from 'react-native';
+
+class Example extends Component {
+  render() {
+    return <Text>Sorry, this demo is not available on web</Text>;
+  }
+}
+
+export const ComboWithGHScroll = Example;
+export const ComboWithRNScroll = Example;
