--- Example/index.ios.js
@@ -1,5 +1,0 @@
-import { AppRegistry } from 'react-native';
-
-import ExampleApp from './App';
-
-AppRegistry.registerComponent('Example', () => ExampleApp);
